Add secure transport TLS module
Secure Transport is a TLS library part of the Security framework
(preinstalled on every iOS and MacOS device). This library does
certificate validation during handshake automatically using the
root certificates from the preinstalled certificate store.

The main reason for this module is proper certificate validation
on iOS devices. This is not possible with gnutls, because there is
no access to the root certificates for external applications.
The module is also intended for use on OSX.