src: darwin/thread: reimplement monotonic clock for darwin
This is based on code which was silently removed in
0f9a94f2183fc6c6e9ce8b77eb497aab160cbbbd and
45c2bc0811cc2d9fe8e11448a05742d7ff6da1d8.

The patch solves timing issues if the user changes the system time, for
instance while playback.

Another timing issue was pointed out if vlc_cond_init_with_daytime is used.
This remains unfixed, and is documented in a Fixme.

close #10072