Qt: fix playlist issues when Qt works as a dialog provider
For the Qt plugin to keep on working both as an interface module and
a dialog/menu/extension provider, the new implementation of pl_Get( p_intf )
can no longer be used blindly, since it now assumes that the playlist is the
parent object. In the latter case, the parent object is the calling interface
(for instance, skins2) and the playlist is actually the grandparent !

As a rule of thumb, THEPL which is now initialized appropriately should be
used instead of pl_Get( p_intf ) throughout Qt.

This fixes trac #10421