Win32: workaround awesome Windows key design
Multimedia Keys, on Windows, generate both WM_APPCOMMAND and WM_KEYDOWN,
even if you accept the first, because that makes so much sense.
But some remotes only generate the APPCOMMAND ones, of course...

Therefore, Qt will get the messages twice and do play/pause.
This is not the correct fix, I'm afraid, but this will do until someone
redoes the configuration of the options and fixes the APPCOMMAND
handling with a correct MCE remote. This will do it for now...

Close #9521