zvbi: Set stop pts only for subtitle teletext
A maximum display time does not make sense for real teletext pages. Instead
they should be displayed until the user disables teletext or the page is
updated. Therefore do not set a stop timestamp on normal pages, but only when
rendering in text mode, which is meant for subtitles, where the timeout makes
sense.

Signed-off-by: Dennis Hamester <dhamester@jusst.de>
Signed-off-by: Julian Scheel <julian@jusst.de>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>