demux/ts.c: Solve 'lost synchro' warnings when using a file input.
Solve 'lost synchro' warnings when using a file input. The warning was
triggered by the functions GetFirstPCR(), CheckPCR() and GetLastPCR().
They searched for a TS packet which contains a PCR but did not take
the TS packet size boundarys into account. By rounding to the nearest
p_sys->i_packet_size this warning was resolved.