transcode: don't care other tracks drift when calculating pts
Also don't care input fps and only calculate what should be the
next output pts. We anyway define output fps to be either 25fps
or input fps if it's available.

This should help to eliminate skips/visual artefacts on video if
there is small skips in input. And output constans fps from transcode.

Fixes #10453 for me