Remove redundant check
We only setup the timer if we are configured to control iTunes, so there should be no reason to check that we control iTunes, before we check if we have setup the timer (only to tear it down).

This change is motivated by using the timer to control (other) third party music players.