mediacodec: do not accept codecs declaring no capabilities
Previously, if profileLevels was NULL or length 0, the codec was
accepted.  Some devices have several codecs for mime type video/avc,
the first one is rejected based on the profile capability but the
second one was accepted since there was no capability declared. Since
the actual capability of the second codec was not better than the
first one, decoding failed.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>