avcodec: allocate more hardware surfaces for threaded decoding
The number of threads impacts the number of surfaces that should be
allocated by the HW acceleration backend. If not enough surfaces are
allocated, the VAAPI and DXVA modules returned the oldest surface
currently in used. This technique can cause many visual glitches with
some samples. The number of allocated surfaces now scales with the
number of threads used by avcodec.

Fix #9887

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>