video output: fix offsets and visible dimensions of filtered video
This parameter copy was working around broken format-changing video
filters, but it broke filters updating the format correctly
(deinterlace, transform and croppadd).

Refs #10745. Note that transform is still not working correctly: the
aspect ratio is broken for transpositation and quarter rotations. This
latent bug should vanish with the video orientation support later.
Otherwise the proper fix is to allow the filters chain to change the
format.