Fix glitching at start of stream playback
This patch fixes a problem that occurs at the start of playback of MPEG2
streams.  The symptom is that an initial good frame will display and
then the video appears to skip back a frame or two and some
macroblocking is observed. The stream then plays correctly. This is
really obvious when switching between SAPed multicast streams, where
you're jumping right in to a lot of movement and probably not starting
with an I frame.

The issue appears to be some code introduced way back in 2004 (commit
eaefb8508dba22fb04193241347a072caac34cf1) which resubmits data to the
video codec. The commit notes suggest it was to avoid dropping the first
I frame, but I see no obviously lost I frames after removing the code
with my test samples. Presumably this was a work-around for a libavcodec
issue that is no longer required.

Therefore, this patch basically just reverts the 2004 commit in its
current form.

Signed-off-by: Ilkka Ollakka <ileoo@videolan.org>