playlist: repair playlist order on shuffle (fix #4857)
Selecting first item to play in random mode as null, the old method
was giving a random item at a random position which resulted in only a
part of the queue playing. If we set first item to NULL, playlist
automaticaly selects first item from the reshuffled queue in the
request processing loop.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>