Qt: fix multiple instances of a singleton if an input is already started
Using THEMIM at this level results in instantiating multiples copies of
what should be a singleton. This is because any preexisting input needs
to be initialized in the constructor of the THEMIN (MainInputManager). This
leads to recursively calling the constructor over and over.

This patch directly calls the underlying functions(no functional change).