Qt: early instantiation for THEMIM
Rather than wait for the first call to THEMIM to instantiate this singleton,
do it explicitly at an early stage of the plugin initialization.

This is useful for Qt as a dialog provider, since otherwise THEMIM doesn't
get a chance to be set up until, e.g. the first call to a popupmenu,
which then makes this menu slightly longer and noticeable to display.

Also, remove the unneeded instantiation in the main_interface. (THEMIM has
already been used a few lines earlier, which makes the call useless).