qt4: Replace redundant virtual with Q_DECL_OVERRIDE
Declaring a virtual function in a subclass as virtual may be a good
documentation that we are overriding a method from the superclass but
has no effect during compilation. With C++11, we can use the override
keyword for this (which will even trigger a compiler error if we try to
override a non-virtual function).

To stay backwards-compatible, we use Qt5's Q_DECL_OVERRIDE macro to
support C++ <11.

Signed-off-by: Tristan Matthews <le.businessman@gmail.com>