mmal: Move codec plugin into hw/mmal module
The mmal codec plugin is built as a part of the mmal module now. This drops
the ability to en-/disable codec support independently from the vout, but in
fact currently there is no mmal implementation which wouldn't support both
modules, so it seems sane to build them together as this unclutters the
configure scripts a little.

Signed-off-by: Julian Scheel <julian@jusst.de>
Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>