mmal: Add deinterlace filter
Add a video_filter module which uses the Broadcom Multi Media Abstraction
Layer (mmal) to deinterlace video in hardware on supported platforms like the
Raspberry Pi.
As of now it is only working in opaque mode, due to a broadcom firmware
limitation.  Thus it can only be used along with mmal/codec and mmal/vout.

Signed-off-by: Julian Scheel <julian@jusst.de>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>