vout: save previous display size regardless of the fullscreen flag
This is simpler and is also more robust. The display no longer needs to
know whether it is fullscreen or not when reporting the size.

This should fix bugs whereby leaving fullscreen forces the window to be
maximized. On the other hand, it will fail if the windowing system
automatically resize a fullscreen window (due to changing view port).