vout: ignore CHANGE_DISPLAY_SIZE failure
In practice, video output display plugins return an error when they do
not support arbitrary sizes. However nowadays, this control is requested
only when the video output or the video window report a change of size.
That cannot be allowed to fail: the display size is what it is
regardless of the result of the control request.