Adds batch convert support to the VLC GUI.
This commit modifies the convert wizard to accept multiple files from the file dialog box

The GUI should operate the same when a single file is selected (providing
an option of where and what to name the file), but when multiple files are
selected the files are placed into the same folder with the same name and a
new extention (there is an option to append -converted if you are converting
to the same extention).
There are some tooltips to explain this operation.

Most of the changes are pretty straight-forward, converting QStrings to
QStringLists and passing the full list of MRLs around. The playlist already
supports the batch processing in a pretty straightforward way, so there's no
issues there.

StandardPanel.cpp was modified to create a temp QStringList for passing to the
streamingDialog, as it now takes the full list of input files rather than just
one

Convert.cpp/hpp modified to take the QStringList and to behave (semi)-intelligently when reciving multiple files,
Open.cpp modified to pass the MRLs rather than MRL[0]
and the Dialogs_provider.cpp/hpp to pass the MRLs along, pretty much.

It also clears the playlist when recieving a list and starts the playlist from the beginning when it's added all the files to be converted.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>