Qt: add QT_STATICPLUGIN define as per the qt spec
The QT spec says we should use QT_STATICPLUGIN defined if building a
static plugin.

As we are using out own makefile and not a qt .pro project we have to
define it ourselves.

This is then used to decide if we should import the modules
QWindowsIntegrationPlugin and AccessibleFactory as these are not needed
on a shared build.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>