Support for LG N1A1 UPnP server
Connecting to a LG N1A1 NAS device gave:
upnp services discovery error: browse() response parsing failed

There is a "ugly hack" ("The DIDL document is extracted from the Result tag,
then wrapped into a valid XML header and a new root tag") in the code now.
This actually break parsing the reply by the LG N1A1.

Now, we try parsing the reply as is, and uses the "ugly hack" as a fallback only.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>