mmal/codec: Enable error_concealment for h264
Until now error concealment was disabled for h264 decoding because the mmal
h264 decoder did not properly parse SEI so that depending on the stream it
could happen that no valid picture would be detected at all and therefore no
picture would be shown at all. Latest raspberry firmware fixed this, so that
error concealment works properly and can be enabled.

Signed-off-by: Julian Scheel <julian@jusst.de>
Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>