libvlc: go to libvlc_Ended rather than libvlc_Stopped at EOS
This is what the tests expect, and was the final status before the
merge of ABORT and DEAD input events.

In fact, libvlc_Stopped does not really correspond to anything: it has
no STOPPED_S equivalent in the input thread state. The only way to
reach it is to call libvlc_media_player_stop() explicitly.
Regardless, lets keep it for backward compatibility.