input: tickless pause
Now we no longer update times and statistics in the control loop
(where it was kinda pointless). If there is no wake-up from the ES
output, then the input thread only needs to wait for control requests
- which means it can sleep without time-out.

In practice, that corresponds to the input thread being paused and
not buffering (buffering while paused is possible due to seek).