input: add b_net variable in item
When an item was a NET/STREAM type, there was no way to distinguish if the item
was a file, a playlist or directory.

Add a new variable, b_net in addition to i_type. This variable is automatically
set when creating a new Item (set to true for ITEM_TYPE_STREAM). It can be
overridden via the new input_item_NewWithTypeExt function. This new function
will allow accesses to create FILE, or DIRECTORY items with b_net set to true.

Unless forced, the preparser won't parse items with b_net == true.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>