video_output: fix controls and events not processed
This issue is easily reproducible with VDPAU activated, with a 60fps ts file,
see #14199.

With this sample, The video_output Thread is looping in ThreadDisplayPicture
way more often than with other videos. Consequently, vout_ManageWrapper and
ThreadControl are not called enough. As a result, subtitles are processed too
late, the mouse isn't able to hide/unhide, we can be stuck in fullscreen mode,
or second click on video to pop up the menu doesn't work.

To fix this issue: don't loop in ThreadDisplayPicture and don't wait in
vout_control_Pop if a picture was previously displayed.

Fixes #14199

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>