input: add support for pointer options to input item
Pointers cannot be inherited as normal configuration item for somewhat
obvious reasons. For output parameters, this is not much of an issue as
LibVLC sets the pointer values directly on its media player object. But
for input-related parameters, LibVLC would need to store pointer values
in the media item. Thus this change to input items.

Note that pointers are considered intrinsically unsafe, so there are no
flags for the time being.